feat: labels attribute support
added support to button, input, meter, output, progress, select and textarea

moved the actual implementation to helper modules
