+++ lib/jsdom/living/helpers/form-controls.js
@@ -18,6 +18,10 @@ const {
 } = require("./dates-and-times");
 const whatwgURL = require("whatwg-url");
 
+const NodeList = require("../generated/NodeList");
+const { domSymbolTree } = require("../helpers/internal-constants");
+const NODE_TYPE = require("../node-type");
+const { getRootNode } = require("../helpers/traversal");
 const submittableLocalNames = new Set(["button", "input", "keygen", "object", "select", "textarea"]);
 
 exports.isDisabled = formControl => {
@@ -58,6 +62,51 @@ exports.normalizeToCRLF = string => {
     .replace(/\r$/, "\r\n")
     .replace(/([^\r])\n/g, "$1\r\n")
     .replace(/^\n/, "\r\n");
+};
+
+exports.isLabelable = node => {
+  // labelable logic defined at: https://html.spec.whatwg.org/multipage/forms.html#category-label
+  if (node.nodeType !== NODE_TYPE.ELEMENT_NODE) {
+    return false;
+  }
+
+  switch (node.tagName) {
+    case "BUTTON":
+    case "KEYGEN":
+    case "METER":
+    case "OUTPUT":
+    case "PROGRESS":
+    case "SELECT":
+    case "TEXTAREA":
+      return true;
+
+    case "INPUT":
+      return node.type !== "hidden";
+  }
+
+  return false;
+};
+
+exports.getLabelsForLabelable = labelable => {
+  if (!exports.isLabelable(labelable)) {
+    return null;
+  }
+  if (!labelable._labels) {
+    const root = getRootNode(labelable);
+    labelable._labels = NodeList.create([], {
+      element: root,
+      query: () => {
+        const nodes = [];
+        for (const descendant of domSymbolTree.treeIterator(root)) {
+          if (descendant.control === labelable) {
+            nodes.push(descendant);
+          }
+        }
+        return nodes;
+      }
+    });
+  }
+  return labelable._labels;
 };
 
 // https://html.spec.whatwg.org/multipage/forms.html#valid-e-mail-address
