+++ lib/jsdom/living/helpers/traversal.js
@@ -82,3 +82,10 @@ exports.firstDescendantWithHTMLLocalName = (parent, localName) => {
   return null;
 };
 
+exports.getRootNode = node => {
+  let root;
+  for (const ancestor of domSymbolTree.ancestorsIterator(node)) {
+    root = ancestor;
+  }
+  return root;
+};
