+++ lib/jsdom/living/nodes/HTMLButtonElement-impl.js
@@ -5,6 +5,7 @@ const { isDisabled } = require("../helpers/form-controls");
 const DefaultConstraintValidationImpl =
   require("../constraint-validation/DefaultConstraintValidation-impl").implementation;
 const { mixin } = require("../../utils");
+const { getLabelsForLabelable } = require("../helpers/form-controls");
 
 class HTMLButtonElementImpl extends HTMLElementImpl {
   constructor(args, privateData) {
@@ -25,6 +26,10 @@ class HTMLButtonElementImpl extends HTMLElementImpl {
   _getValue() {
     const valueAttr = this.getAttribute("value");
     return valueAttr === null ? "" : valueAttr;
+  }
+
+  get labels() {
+    return getLabelsForLabelable(this);
   }
 
   get form() {
