+++ lib/jsdom/living/nodes/HTMLButtonElement.webidl
@@ -20,5 +20,5 @@ interface HTMLButtonElement : HTMLElement {
   boolean reportValidity();
   void setCustomValidity(DOMString error);
 
-//  readonly attribute NodeList labels;
+  readonly attribute NodeList labels;
 };
