+++ lib/jsdom/living/nodes/HTMLInputElement-impl.js
@@ -4,7 +4,6 @@ const DOMException = require("domexception");
 const Event = require("../generated/Event");
 const FileList = require("../generated/FileList");
 const HTMLElementImpl = require("./HTMLElement-impl").implementation;
-const NodeList = require("../generated/NodeList");
 const idlUtils = require("../generated/utils");
 const DefaultConstraintValidationImpl =
   require("../constraint-validation/DefaultConstraintValidation-impl").implementation;
@@ -12,6 +11,7 @@ const ValidityState = require("../generated/ValidityState");
 const { mixin } = require("../../utils");
 const { domSymbolTree, cloningSteps } = require("../helpers/internal-constants");
 const { closest } = require("../helpers/traversal");
+const { getLabelsForLabelable } = require("../helpers/form-controls");
 const {
   isDisabled,
   isValidEmailAddress,
@@ -152,14 +152,6 @@ const convertStringToNumberByTypeMap = new Map([
   // https://html.spec.whatwg.org/multipage/input.html#range-state-(type=range):concept-input-value-string-number
   ["range", parseFloatingPointNumber]
 ]);
-
-function getRootNode(node) {
-  let root;
-  for (const ancestor of domSymbolTree.ancestorsIterator(node)) {
-    root = ancestor;
-  }
-  return root;
-}
 
 class HTMLInputElementImpl extends HTMLElementImpl {
   constructor(args, privateData) {
@@ -328,30 +320,8 @@ class HTMLInputElementImpl extends HTMLElementImpl {
     return this._otherRadioGroupElements.some(radioGroupElement => radioGroupElement.checked);
   }
 
-  _isHiddenInput() {
-    return this.tagName === "INPUT" && this.type === "hidden";
-  }
-
   get labels() {
-    if (this._isHiddenInput()) {
-      return null;
-    }
-    if (!this._labels) {
-      const root = getRootNode(this);
-      this._labels = NodeList.create([], {
-        element: root,
-        query: () => {
-          const nodes = [];
-          for (const descendant of domSymbolTree.treeIterator(root)) {
-            if (descendant.control === this) {
-              nodes.push(descendant);
-            }
-          }
-          return nodes;
-        }
-      });
-    }
-    return this._labels;
+    return getLabelsForLabelable(this);
   }
 
   get form() {
