+++ lib/jsdom/living/nodes/HTMLLabelElement-impl.js
@@ -4,30 +4,7 @@ const HTMLElementImpl = require("./HTMLElement-impl").implementation;
 const MouseEvent = require("../generated/MouseEvent");
 const { domSymbolTree } = require("../helpers/internal-constants");
 const NODE_TYPE = require("../node-type");
-const { isDisabled } = require("../helpers/form-controls");
-
-function isLabelable(node) {
-  // labelable logic defined at: https://html.spec.whatwg.org/multipage/forms.html#category-label
-  if (node.nodeType !== NODE_TYPE.ELEMENT_NODE) {
-    return false;
-  }
-
-  switch (node.tagName) {
-    case "BUTTON":
-    case "KEYGEN":
-    case "METER":
-    case "OUTPUT":
-    case "PROGRESS":
-    case "SELECT":
-    case "TEXTAREA":
-      return true;
-
-    case "INPUT":
-      return node.type !== "hidden";
-  }
-
-  return false;
-}
+const { isLabelable, isDisabled } = require("../helpers/form-controls");
 
 function getRootNode(node) {
   let root;
