+++ lib/jsdom/living/nodes/HTMLMeterElement-impl.js
@@ -2,6 +2,7 @@
 
 const HTMLElementImpl = require("./HTMLElement-impl").implementation;
 const { parseFloatingPointNumber } = require("../helpers/strings");
+const { getLabelsForLabelable } = require("../helpers/form-controls");
 
 class HTMLMeterElementImpl extends HTMLElementImpl {
   // https://html.spec.whatwg.org/multipage/form-elements.html#concept-meter-minimum
@@ -115,6 +116,10 @@ class HTMLMeterElementImpl extends HTMLElementImpl {
     }
 
     return candidate > maximumValue ? maximumValue : candidate;
+  }
+
+  get labels() {
+    return getLabelsForLabelable(this);
   }
 
   get value() {
