+++ lib/jsdom/living/nodes/HTMLOutputElement-impl.js
@@ -4,10 +4,15 @@ const HTMLElementImpl = require("./HTMLElement-impl").implementation;
 const DefaultConstraintValidationImpl =
   require("../constraint-validation/DefaultConstraintValidation-impl").implementation;
 const { mixin } = require("../../utils");
+const { getLabelsForLabelable } = require("../helpers/form-controls");
 
 class HTMLOutputElementImpl extends HTMLElementImpl {
   _barredFromConstraintValidationSpecialization() {
     return true;
+  }
+
+  get labels() {
+    return getLabelsForLabelable(this);
   }
 }
