+++ lib/jsdom/living/nodes/HTMLOutputElement.webidl
@@ -16,5 +16,5 @@ interface HTMLOutputElement : HTMLElement {
   boolean reportValidity();
   void setCustomValidity(DOMString error);
 
-//  readonly attribute NodeList labels;
+  readonly attribute NodeList labels;
 };
