+++ lib/jsdom/living/nodes/HTMLProgressElement-impl.js
@@ -1,8 +1,13 @@
 "use strict";
 
 const HTMLElementImpl = require("./HTMLElement-impl").implementation;
+const { getLabelsForLabelable } = require("../helpers/form-controls");
 
-class HTMLProgressElementImpl extends HTMLElementImpl { }
+class HTMLProgressElementImpl extends HTMLElementImpl {
+  get labels() {
+    return getLabelsForLabelable(this);
+  }
+}
 
 module.exports = {
   implementation: HTMLProgressElementImpl
