+++ lib/jsdom/living/nodes/HTMLProgressElement.webidl
@@ -4,5 +4,5 @@ interface HTMLProgressElement : HTMLElement {
 //  [CEReactions] attribute double value;
 //  [CEReactions] attribute double max;
 //  readonly attribute double position;
-//  readonly attribute NodeList labels;
+  readonly attribute NodeList labels;
 };
