+++ lib/jsdom/living/nodes/HTMLSelectElement-impl.js
@@ -13,6 +13,7 @@ const HTMLCollection = require("../generated/HTMLCollection");
 const HTMLOptionsCollection = require("../generated/HTMLOptionsCollection");
 const { domSymbolTree } = require("../helpers/internal-constants");
 const { closest } = require("../helpers/traversal");
+const { getLabelsForLabelable } = require("../helpers/form-controls");
 
 class HTMLSelectElementImpl extends HTMLElementImpl {
   constructor(args, privateData) {
@@ -63,8 +64,8 @@ class HTMLSelectElementImpl extends HTMLElementImpl {
         let disabled = option.hasAttribute("disabled");
         const parentNode = domSymbolTree.parent(option);
         if (parentNode &&
-            parentNode.nodeName.toUpperCase() === "OPTGROUP" &&
-            parentNode.hasAttribute("disabled")) {
+          parentNode.nodeName.toUpperCase() === "OPTGROUP" &&
+          parentNode.hasAttribute("disabled")) {
           disabled = true;
         }
 
@@ -144,6 +145,10 @@ class HTMLSelectElementImpl extends HTMLElementImpl {
     for (let i = 0; i < this.options.length; i++) {
       this.options.item(i).selected = i === index;
     }
+  }
+
+  get labels() {
+    return getLabelsForLabelable(this);
   }
 
   get value() {
