+++ lib/jsdom/living/nodes/HTMLSelectElement.webidl
@@ -32,5 +32,5 @@ interface HTMLSelectElement : HTMLElement {
   boolean reportValidity();
   void setCustomValidity(DOMString error);
 
-//  readonly attribute NodeList labels;
+  readonly attribute NodeList labels;
 };
