+++ lib/jsdom/living/nodes/HTMLTextAreaElement-impl.js
@@ -9,7 +9,7 @@ const { mixin } = require("../../utils");
 
 const DOMException = require("domexception");
 const { closest } = require("../helpers/traversal");
-const { normalizeToCRLF } = require("../helpers/form-controls");
+const { normalizeToCRLF, getLabelsForLabelable } = require("../helpers/form-controls");
 const { childTextContent } = require("../helpers/text");
 
 class HTMLTextAreaElementImpl extends HTMLElementImpl {
@@ -42,6 +42,10 @@ class HTMLTextAreaElementImpl extends HTMLElementImpl {
     }
   }
 
+  get labels() {
+    return getLabelsForLabelable(this);
+  }
+
   get form() {
     return closest(this, "form");
   }
@@ -70,6 +74,7 @@ class HTMLTextAreaElementImpl extends HTMLElementImpl {
   get textLength() {
     return this.value.length; // code unit length (16 bit)
   }
+
   get type() {
     return "textarea";
   }
@@ -79,39 +84,49 @@ class HTMLTextAreaElementImpl extends HTMLElementImpl {
     event.initEvent("select", true, true);
     this.dispatchEvent(event);
   }
+
   _getValueLength() {
     return typeof this.value === "string" ? this.value.length : 0;
   }
+
   select() {
     this._selectionStart = 0;
     this._selectionEnd = this._getValueLength();
     this._selectionDirection = "none";
     this._dispatchSelectEvent();
   }
+
   get selectionStart() {
     return this._selectionStart;
   }
+
   set selectionStart(start) {
     this.setSelectionRange(start, Math.max(start, this._selectionEnd), this._selectionDirection);
   }
+
   get selectionEnd() {
     return this._selectionEnd;
   }
+
   set selectionEnd(end) {
     this.setSelectionRange(this._selectionStart, end, this._selectionDirection);
   }
+
   get selectionDirection() {
     return this._selectionDirection;
   }
+
   set selectionDirection(dir) {
     this.setSelectionRange(this._selectionStart, this._selectionEnd, dir);
   }
+
   setSelectionRange(start, end, dir) {
     this._selectionEnd = Math.min(end, this._getValueLength());
     this._selectionStart = Math.min(start, this._selectionEnd);
     this._selectionDirection = dir === "forward" || dir === "backward" ? dir : "none";
     this._dispatchSelectEvent();
   }
+
   setRangeText(repl, start, end, selectionMode = "preserve") {
     if (arguments.length < 2) {
       start = this._selectionStart;
