+++ lib/jsdom/living/nodes/HTMLTextAreaElement.webidl
@@ -29,7 +29,7 @@ interface HTMLTextAreaElement : HTMLElement {
   boolean reportValidity();
   void setCustomValidity(DOMString error);
 
-//  readonly attribute NodeList labels;
+  readonly attribute NodeList labels;
 
   void select();
   attribute unsigned long selectionStart;
